[#9] Separate Customer Model and Logic from Account App
