+++ account/models.py
@@ -5,6 +5,7 @@ from django.contrib.auth.models import PermissionsMixin
 
 class User(AbstractBaseUser, PermissionsMixin):
     email = models.EmailField(unique=True)
+    name = models.CharField(max_length=50)
     is_active = models.BooleanField(default=False)
     is_staff = models.BooleanField(default=False)
     is_superuser = models.BooleanField(default=False)
@@ -24,55 +25,3 @@ class User(AbstractBaseUser, PermissionsMixin):
     
     def has_module_perms(self, app_label):
         return self.is_superuser
-
-
-
-
-
-STATE_CHOICES = (
-   ("Andaman and Nicobar Islands","Andaman and Nicobar Islands"),
-   ("Andhra Pradesh","Andhra Pradesh"),
-   ("Arunachal Pradesh","Arunachal Pradesh"),
-   ("Assam","Assam"),
-   ("Bihar","Bihar"),
-   ("Chhattisgarh","Chhattisgarh"),
-   ("Chandigarh","Chandigarh"),
-   ("Dadra and Nagar Haveli","Dadra and Nagar Haveli"),
-   ("Daman and Diu","Daman and Diu"),
-   ("Delhi","Delhi"),
-   ("Goa","Goa"),
-   ("Gujarat","Gujarat"),
-   ("Haryana","Haryana"),
-   ("Himachal Pradesh","Himachal Pradesh"),
-   ("Jammu and Kashmir","Jammu and Kashmir"),
-   ("Jharkhand","Jharkhand"),
-   ("Karnataka","Karnataka"),
-   ("Kerala","Kerala"),
-   ("Ladakh","Ladakh"),
-   ("Lakshadweep","Lakshadweep"),
-   ("Madhya Pradesh","Madhya Pradesh"),
-   ("Maharashtra","Maharashtra"),
-   ("Manipur","Manipur"),
-   ("Meghalaya","Meghalaya"),
-   ("Mizoram","Mizoram"),
-   ("Nagaland","Nagaland"),
-   ("Odisha","Odisha"),
-   ("Punjab","Punjab"),
-   ("Pondicherry","Pondicherry"),
-   ("Rajasthan","Rajasthan"),
-   ("Sikkim","Sikkim"),
-   ("Tamil Nadu","Tamil Nadu"),
-   ("Telangana","Telangana"),
-   ("Tripura","Tripura"),
-   ("Uttar Pradesh","Uttar Pradesh"),
-   ("Uttarakhand","Uttarakhand"),
-   ("West Bengal","West Bengal")
-)
-
-class Customer(models.Model):
-    user = models.ForeignKey(User, on_delete=models.CASCADE)
-    name = models.CharField(max_length=100)
-    locality = models.CharField(max_length=200)
-    city = models.CharField(max_length=50)
-    zipcode = models.IntegerField()
-    state = models.CharField(choices=STATE_CHOICES, max_length=50)
