+++ account/templates/account/password_change.html
@@ -6,7 +6,7 @@
 
 <div class="row">
   <div class="col-md-4">
-    {% include 'account/components/profile_sidebar.html' %}
+    {% include 'customer/components/sidebar.html' %}
   </div>
   <div class="col-md-8">
       <h3 class="mb-3">Change password</h3>
