+++ core/templates/core/components/navbar.html
@@ -7,7 +7,11 @@
 
         <div class="collapse navbar-collapse" id="navbarSupportedContent">
             <ul class="navbar-nav me-auto">
-                
+                {% if request.user.is_authenticated %}
+                    <li class="nav-item">
+                        <a class="nav-link text-white" href="{% url 'customer_dashboard' %}">Dashboard</a>
+                    </li>
+                {% endif %}
                 <li class="nav-item dropdown">
                     <a class="nav-link text-white dropdown-toggle" href="#" id="navbarDropdown" role="button" data-bs-toggle="dropdown" aria-expanded="false">
                         Electronics
@@ -28,16 +32,7 @@
                 </li>
                 {% if request.user.is_authenticated %}
                     <li class="nav-item">
-                        <a class="nav-link text-white" href="{% url 'dashboard' %}">Dashboard</a>
-                    </li>
-                    <li class="nav-item">
                         <a class="nav-link text-white" href="#">Carts <span class="badge bg-danger" id="badge">3</span></a>
-                    </li>
-                    <li  class="nav-item">
-                        <form action="{% url 'logout' %}" method="post">
-                            {% csrf_token %}
-                            <button type="submit" class="nav-link text-white">Logout</button>
-                        </form>
                     </li>
                 {% else %}
                     <li class="nav-item">
