+++ customer/models.py
@@ -1,0 +1,50 @@
+from django.db import models
+from account.models import User
+
+STATE_CHOICES = (
+   ("Andaman and Nicobar Islands","Andaman and Nicobar Islands"),
+   ("Andhra Pradesh","Andhra Pradesh"),
+   ("Arunachal Pradesh","Arunachal Pradesh"),
+   ("Assam","Assam"),
+   ("Bihar","Bihar"),
+   ("Chhattisgarh","Chhattisgarh"),
+   ("Chandigarh","Chandigarh"),
+   ("Dadra and Nagar Haveli","Dadra and Nagar Haveli"),
+   ("Daman and Diu","Daman and Diu"),
+   ("Delhi","Delhi"),
+   ("Goa","Goa"),
+   ("Gujarat","Gujarat"),
+   ("Haryana","Haryana"),
+   ("Himachal Pradesh","Himachal Pradesh"),
+   ("Jammu and Kashmir","Jammu and Kashmir"),
+   ("Jharkhand","Jharkhand"),
+   ("Karnataka","Karnataka"),
+   ("Kerala","Kerala"),
+   ("Ladakh","Ladakh"),
+   ("Lakshadweep","Lakshadweep"),
+   ("Madhya Pradesh","Madhya Pradesh"),
+   ("Maharashtra","Maharashtra"),
+   ("Manipur","Manipur"),
+   ("Meghalaya","Meghalaya"),
+   ("Mizoram","Mizoram"),
+   ("Nagaland","Nagaland"),
+   ("Odisha","Odisha"),
+   ("Punjab","Punjab"),
+   ("Pondicherry","Pondicherry"),
+   ("Rajasthan","Rajasthan"),
+   ("Sikkim","Sikkim"),
+   ("Tamil Nadu","Tamil Nadu"),
+   ("Telangana","Telangana"),
+   ("Tripura","Tripura"),
+   ("Uttar Pradesh","Uttar Pradesh"),
+   ("Uttarakhand","Uttarakhand"),
+   ("West Bengal","West Bengal")
+)
+
+class Customer(models.Model):
+    user = models.ForeignKey(User, on_delete=models.CASCADE)
+    customer_name = models.CharField(max_length=100)
+    locality = models.CharField(max_length=200)
+    city = models.CharField(max_length=50)
+    zipcode = models.IntegerField()
+    state = models.CharField(choices=STATE_CHOICES, max_length=50)
