+++ customer/templates/customer/components/form.html
@@ -1,0 +1,16 @@
+{% if form.non_field_errors %}
+    {% for error in form.non_field_errors %}
+        <p>{{error}}</p>
+    {% endfor %}
+{% endif %}
+
+
+
+{% for field in form %}
+    <div class="form-group mb-3">
+        {{field.label_tag}} {{field}}
+        {% for error in field.errors %}
+            <small class="text-danger">{{error}}</small>
+        {% endfor %}
+    </div>
+{% endfor %}
