+++ customer/templates/customer/components/sidebar.html
@@ -1,24 +1,22 @@
-{% load custom_tags %}
-{% get_username as username %}
-
-
-
 <div class="bg-light p-3 rounded mb-3">
     <h5 class="lead mb-3">
-        Welcome, {% if username %} {{username}} {% endif %}
+        Welcome, {{ request.user.name }}
     </h5>
     <ul>
         <li class="my-2">
-            <a href="{% url 'dashboard' %}">Profile</a>
-        </li>
-        <li  class="my-2">
-            <a href="{% url 'address' %}">Address</a>
+            <a href="{% url 'customer_dashboard' %}">Dashboard</a>
         </li>
         <li  class="my-2">
             <a class="{% if active %} {% endif %}" href="#">Orders</a>
         </li>
         <li  class="my-2">
             <a href="{% url 'password_change' %}">Change password</a>
+        </li>
+        <li  class="my-2">
+            <form action="{% url 'logout' %}" method="post">
+                {% csrf_token %}
+                <button type="submit" class="btn btn-link p-0">Logout</button>
+            </form>
         </li>
     </ul>
 </div>
