+++ src/urls.py
@@ -21,4 +21,5 @@ urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('core.urls')),
     path('account/', include('account.urls')),
+    path('customer/', include('customer.urls')),
 ]
